rename: getParam to getParams
