+++ amplify/backend/function/qaSSMtest/src/index.js
@@ -1,4 +1,4 @@
-const { getParam } = require('/opt/nodejs/getParam');
+const { getParam } = require('/opt/nodejs/getParams');
 
 /**
  * @type {import('@types/aws-lambda').APIGatewayProxyHandler}
@@ -12,6 +12,13 @@ exports.handler = async () => {
   const chainNetwork = await getParam('chainNetwork');
   const appSyncApi = await getParam('appSyncApi');
 
- console.log({graphqlUrl, reputationEndpoint, chainNetworkContract, chainRpcEndpoint, chainNetwork, appSyncApi})
+  console.log({
+    graphqlUrl,
+    reputationEndpoint,
+    chainNetworkContract,
+    chainRpcEndpoint,
+    chainNetwork,
+    appSyncApi,
+  });
   return null;
 };
